Update patch set 2
Patch Set 2:

PS2 adds missing BEGIN/ENDs to demarcate where the code differs from dlmalloc. The use of char for a boolean is switched to an int (to be consistent with other boolean uses like mspace_track_large_chunks). A minor tidy up is done to the if statement.

Patch-set: 2
Label: Code-Review=0
Label: Verified=0