Update patch set 5
Patch Set 5:

One approach to take is forking a child process early, before threads start up. This child process can be used to safely fork additional sandbox processes for the compiler whenever you need them. The parent delegates to that child whenever it needs to start a new compiler job.

This route may seem a little indirect, but it is far more future proof since you can guarantee that it is safe to fork and run before threads come into the picture. Trying to fix things with pthread_atfork() exposes you to deadlock hazards whenever new synchronized code paths are added (e.g. from new libraries and future development) that both the parent and child make use of -- most developers are not going to expect to have to be fork safe in multithreaded code.

Patch-set: 5