Update patch set 4
Patch Set 4:

The workaround is only for ARM. There should be no MIPS shared libraries in the wild that resolve atexit() externally.

This means the atexit and __dso_handle LOCAL symbols you are seeing are the ones in the crt startup files, which is expected. The only target that should contain GLOBAL versions of those symbols is ARM's libc.so

Patch-set: 4