Update patch set 1
Patch Set 1:

I have tested that the compilation fails even on ARM with this patch. The problems seems to be that new headers need to be brought in for 
original/linux/in.h
original/linux/un.h

Also with the below commit to external/kernel-headers, the macros like AF_UNIX and AF_LOCAL are not carried to bionic kernel headers when the update_all.py script is run. The reason being that the new socket.h dropped checks for __GLIBC__.

-#if defined(__KERNEL__) || !defined(__GLIBC__) || (__GLIBC__ < 2)
+#ifdef __KERNEL__


commit 5200d181a297324c676df038f8259b48a2af1176
Author: Yegor Yefremov <yegorslists@googlemail.com>
Date:   Sat Apr 7 23:12:24 2012 +0200

    kernel-headers: update socket.h headers
    
    update/add socket.h headers to comply with kernel v3.3
    
    Change-Id: I41a8acb19e1a4e79d5f4726cda011d056abec965
    Signed-off-by: Yegor Yefremov <yegorslists@googlemail.com>

:100644 100644 5903c5c... d0e77f6... M  original/linux/socket.h

Patch-set: 1