Update patch set 1
Patch Set 1: Looks good to me, approved

the high-order bits (which is what i think you mean in the check-in comment) were non-zero in that bug. this was obvious from strace and led me right to the bug. the trouble was that the kernel guy who looked at this bug didn't use strace --- he added logging to his kernel, presumably _after_ the point at which the kernel does the masking itself.

so i think this would have caught the bug.

Patch-set: 1
Label: Code-Review=+2