Update patch set 1
Patch Set 1:

out of interest, do you know which of the three syscalls is failing for you, and with what errno value? i just wonder whether the failures are anything to worry about. are you testing on an emulator or a real device?

also, i assume you guys are aware that glibc uses the same value for every thread, and want to pay the little extra to have a different value for each thread instead?

Patch-set: 1