libc: Add pthread_barrier_t and related functions.
This adds a simple pthread_barrier_t implementation that satisfies
the Posix requirements + a related unit test.

Advertise the feature through a new macro (__BIONIC_HAVE_PTHREAD_BARRIER_T)
to signal it to client code, to ease porting.

A typical client library that uses this is Google Breakpad, though it
currently comes with a custom, but very similar, implementation of
these at the moment.

Change-Id: I984dc26951e07a2e21b73707e8a13716aeb4bc14