Update patch set 1
Patch Set 1:

thanks for the comments.

so i'm tempted for now to just automatically turn on unaligned access if we have neon. if that works out, i'll come back and remove the extra #define from the code. if it doesn't, i'll come back and add an extra makefile flag so it can be turned off for individual devices (but switch the sense so it's still on by default).

i see that currently nothing in google's tree sets ARCH_ARM_HAVE_32_BYTE_CACHE_LINES. but you're saying that at least Cortex A9 devices should be using it? the internet seems to agree, and says that A8 was 64. is there a table somewhere? do we have the default the wrong way round here?

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=0