Update patch set 1
Patch Set 1:

I think you should give in to temptation in this case.

As far as I know, CA8 and CA15 runs 64B/cache line, CA9 runs 32B/cache line.

I haven't been able to find a single good page on this (including ARM's own site, and Wikipedia), but I have found a set of pages with a lot of info on the Cortex CPUs and variants.

http://www.7-cpu.com/cpu/Cortex-A8.html
http://www.7-cpu.com/cpu/Cortex-A9.html
http://www.7-cpu.com/cpu/Cortex-A15.html

Patch-set: 1