Update patch set 2
Patch Set 2:

According to the tgkill(2) manual:

tgkill()  sends  the  signal  sig to the thread with the thread ID tid in the thread group tgid.  (By contrast,
kill(2) can only be used to send a signal to a process (i.e., thread group) as a whole, and the signal will  be
delivered to an arbitrary thread within that process.)

This showed up in an LTP test. The test had 2 threads. One thread set the signal mask and raised the same signal assuming that the signal handler would not be run. 
Instead the signal handler ran in the other thread leading to random failures.

Patch-set: 2