Aprof : Android Profiler
It's a profiling tool for Android native code; aprof is not only
another gprof implement on Android but also support for profiling
shared library. The most important thing is that aport supports
the profiling for JNI code!
The capability of aprof is similar to what gprof does, it provides
call graph and time sampling profiling, but it's incompatible with
gprof since the gprof can't profile shared library, it's limited by
its representation.

Change-Id: Iddacac440e61d625a4f1165bfa210cd49eaaf3eb