Update patch set 1
Patch Set 1:

Thanks for fixing the MIPS build with the below patch.

https://android-review.googlesource.com/#/c/49909/

However there needs to be another patch as signalfd on MIPS does not work with the header sys/signalfd.h. The definition for off_t needs to be fixed.

The problem is that MIPS kernel headers fcntl.h and stat.h
use off_t and the connecting macro for off_t and 
__kernel_off_t is under "ifndef __KERNEL_STRICT_NAMES". 
__KERNEL_STRICT_NAMES gets defined in libc/tools
/defaults.py. 

The real problem is that most architectures moved towards 
using __kernel_off_t instead of off_t. We plan to do the 
same for MIPS kernel headers too. For the signalfd to work 
as per manpage, I would like to introduce a patch to do the 
token replacement of off_t with __kernel_off_t on MIPS.

Patch-set: 1
Reviewer: Gerrit User 1003296 <1003296@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0