Update patch set 1
Patch Set 1:

I see two separate problems here.
The first problem is that there is a lot of build flags/defines missing in the makefile, so the optimized version is never actually run.
NEON_UNALIGNED_ACCESS, NEON_MEMCPY_ALIGNMENT_DIVIDER (224), and NEON_MEMSET_DIVIDER (132) is supposed to be set.

The last alternative discussed was to not set 'unaligned access', but instead delete the first half of the code (line 52-172).

The second problem is that this patch will disable the performance increase for A9 and restore the function to a state of being optimized for Snapdragon/Krait only, since they have improved memory access for Neon. Would it be possible to disable the ARM-version only for Krait, and possibly for A15?

I haven't measured on a real A15, but I will take some time and measure on a Galaxy Nexus (A9) and Nexus 4 (Krait) soon.

Patch-set: 1