Fix the wrong length of the property ro.build.finger
This patch fixed the wrong size of the property string: ro.build.finger.
The length of the string is 93, but the max length of this string in the system is defined to 92.
The system loads this property from configure file and stored it on shared memory.
The memory of this property will be overlapped after storing the next property.
Therefore, this property cannot be read from shared memory correctly.

Signed-off-by: Panfeng Chang <panfengx.chang@intel.com>
Signed-off-by: Jackie Wu <jackie.wu@intel.com>
Signed-off-by: Jun Tian <jun.j.tian@intel.com>
Signed-off-by: Chen Liang <liang.l.chen@intel.com>
Signed-off-by: Zhu Yanglin <zhu.yanglinx@intel.com>

Change-Id: I4831880ca49bafcdc9d936f246654fd4de7e5126