Update patch set 3
Patch Set 3:

The test in question is for sigaction and can be seen here:

http://ltp.git.sourceforge.net/git/gitweb.cgi?p=ltp/ltp.git;a=blob;f=testcases/open_posix_testsuite/conformance/interfaces/sigaction/22-26.c;h=d0eadfa585f118ad6bddcaadadddd73d7808db91;hb=HEAD

In the normal Linux environment this test is built without -lpthread and raise() will use kill() to send signals. When it's built with bionic for Android raise() uses pthread_kill() which causes the deadlock

Yes, using pthreads functions in signal handlers is undefined so technically the test is broken when built for Android, but the program itself is not pthreads aware and I think it would be a good idea to have a workaround for naive programs.

Patch-set: 3