Update patch set 1
Patch Set 1: Code-Review-2

this is the wrong way to do it. the whole point of the vdso stuff is that the dynamic linker can just add it to its list of loaded shared objects, and then everything else just works. we should do that rather than hack the places that iterate over the list.

Patch-set: 1
Label: Code-Review=-2