Update patch set 2
Patch Set 2:

I'll get to working on this in the next week.  For now I found an interesting related aspect to this bug.  I recently ported some of my modifications of the DalvikVM and my internal daemons prevented the zygote proceeding past dalvik.system.Zygote.waitUntilAllThreadsStopped because it looks at /proc/self/task and sees my daemons (which I tear down a little later).  Looking through the code it looks like there is no reason for this code to exist as the daemons are already joined.  I strongly suspect that this code was introduced as a pragmatic fix to the issue this patch is trying to fix.

For reference:

    private static void waitUntilAllThreadsStopped() {
        File tasks = new File("/proc/self/task");
        while (tasks.list().length > 1) {
            try {
                // Experimentally, booting and playing about with a stingray, I never saw us
                // go round this loop more than once with a 10ms sleep.
                Thread.sleep(10);
            } catch (InterruptedException ignored) {
            }
        }
    }

Patch-set: 2