Update patch set 4
Patch Set 4:

I think it would be useful to have a test case of reaching but not exceeding the end of the buffer to make sure there is not an off by one error, paired with one where if it you go over by one you can confirm it detects the overflow. sorry I wasn't clearer before.

Patch-set: 4
Label: Verified=0