Update patch set 4
Patch Set 4:

Finally sucked it up and built the AOSP emulator and ran the bionic tests.  I actually did make a mistake and dropped the code that set the return_value so I guess that was worth it.  Added a test to check the multijoin behavior, I'm not sure of anyway to test the spurious wakeup behavior without peaking into the internal pthread structures and signaling the join_cond when the thread hasn't exited.

Anyway one of the bionic tests still failed, but it seems unlikely to be related.

[==========] 393 tests from 26 test cases ran. (73229 ms total)
[  PASSED  ] 392 tests.
[  FAILED  ] 1 test, listed below:
[  FAILED  ] dlfcn.dlopen_library_with_only_gnu_hash

Patch-set: 4
Reviewer: Gerrit User 1006468 <1006468@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0