Update patch set 1
Patch Set 1:

Could the 'size multiple of 32' performance drop have anything to do with the cache prefetch? Since the Cortex A9 has 32 bytes cache line, I prefetched each line separately in the original code i submitted. In the current AOSP version you're assuming the 'double line fill' is enabled in the L2 cache. Is it, on the nexus for example?

Patch-set: 1