Update patch set 1
Patch Set 1:

So I did a quick optimization, and my version only loses significantly for sizes [16, 31]. It is slightly slower for some of the values between 32 and 60, but it's close. Then for everything else it pulls away pretty quickly.

Here's the crude first version I came up with:

https://android-review.googlesource.com/#/c/60353/

It wins in all cases against the current code.

It does appear that the neon is not a winner except in some narrow cases > 32 <= 60, most likely due to the extra registers that have to be saved/restored for the arm only case. The 16 to 32 case is likely because this code has a quick path for all sizes < 32, while mine sticks with a quick path only for < 16. It's probably worth adding a quick path for < 32 since it only shaves a hair off the < 16 cases, and vastly improves the [16,32) cases.

For all of this, I've been assuming that the normal case is that the addresses being memcopied are aligned to 8 bytes, but if that's not true please let me know.

It would be useful if you happen to have any data on the frequency of memcopy sizes chosen to see if there is any size that would be especially useful to optimize.

Patch-set: 1