Update patch set 1
Patch Set 1:

(1 comment)

Can you point to the exact location of the benchmarks in the link? It's not immediately obvious.

I did some quick benchmarks and it appears that this code is much faster for data in the cache, but isn't so good for data not in the cache.

For example, doing larger copies, this new code is worse than the current code. At sizes about 20,000 you already start to see performance drop off, which is why it looks like there might be issues when the data isn't already in the cache.

Also, I did some quick alignment checking, and those numbers look a little worse even for some of the smaller copies.

Patch-set: 1