Update patch set 1
Patch Set 1:

Since this is exactly the same as the cortex-a9 implementation, then you don't need to have both versions.

You can modify the bionic/libc/arch-arm/cortex-a15/cortex-a15.mk and change the line about the memcpy to use the cortexa9 version.

This code also appears to suffer from the same problem as the cortexa9 code, it starts to lose when the data is not in the cache. Do you have any benchmarks that show this code when the data isn't already in the cache?

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0