Update patch set 1
Patch Set 1:

The problem is that your harness.c code isn't properly initializing the src buffer, so you're getting radically incorrect benchmark scores for copies > 16K.

You need to modify this code:

srandom(1539);

for (int i = 0; i < 16*1024; i++)

Change the 16 to 1024 so the whole buffer is initialized. I would have expected the calloc to force initialization, but apparently not. If you don't do this init, then it appears that something somewhere knows that it's copying memory in an unknown state, and it doesn't actually do the copy.

This also means that any data you have for strcpy/strcmp > 16K is also wrong because there is always a string terminator at the 16K mark. However, I would hope no one would be using strxxx routines on strings of this size.

After I did this, the numbers I've been gathering and the numbers from this tool are much closer (nearly the same actually).

For android, I've been using the micro_bench tool (found in system/extras/micro_bench), which I've been modifying to display better data such as producing a std deviation across a number of different runs.

Patch-set: 1