Update patch set 1
Patch Set 1:

Thanks for looking at this, that's a really useful insight. calloc in glibc (for large sizes at least) may end up relying on mmap zeroing the memory. The kernel does this by mapping the zero page, so in that case a physically tagged cache would only use a page size of cache and and never miss on the src. I'll rework our benchmark framework and re-run the tests.

Patch-set: 1