Update patch set 1
Patch Set 1: Code-Review-2

This needs to be reworked to follow the new framework that supports separate implementations of these low-level routines per processor.

And with all of these changes, please describe how you tested and how you benchmarked the changes and on what platforms you did the benchmarks.

Be aware, there is already a pending change for the memcpy cortexa9 code that is also from Linaro so maybe this change needs to be abandoned.

That change is:

https://android-review.googlesource.com/#/c/60490/1

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2