Update patch set 2
Patch Set 2: Code-Review+2

good. i didn't really understand your logic.

i do think we need to do something about the duplication, but skipping tests is probably not the answer.

i wonder whether we can have something we #include that contains the bodies of all the tests. all we need to be able to eventually get this into CTS is that we can scan the source files and work out all the test names. so as long as we duplicate the function declarations, we don't actually need to duplicate the bodies.

Patch-set: 2
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2