Update patch set 1
Patch Set 1:

I wrote a new benchmark to try and get the uncached performance of memcpy. It's not perfect but it does show that this code is much worse for uncached data (as the sizes get higher, the speed of this algorithm is about half the current one). The cross-over point is at sizes of about 330.

I did this measurement on a nexus 7.

If you want, you can abandon this change and then submit this code as bionic/libc/arch-arm/bionic/memcpy.a9.S. I'll take that code and tweak it to fit into our environment and to tweak it so the small deficiencies are removed. I can send you a note when it's all done and to let you look to see what I did. Let me know what you'd like to do.

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0