Update patch set 1
Patch Set 1:

will.newton: could you upload a patch that separates the copyright notice from the documentation please? our open source/copyright guy says we shouldn't do that ourselves because as long as they're in the same comment, the documentation counts as part of the copyright, which we're not allowed to change. thanks!

Patch-set: 1