Update patch set 2
Patch Set 2:

Indeed, will wait for Ying's response about proper location.  The motive here is that x86 toolchain will become multilib pretty soon, and the default (in compiler) is usually the widest, eg. 64-bit.  We need something in AOSP for 32-bit, and easy way for partner to switch to 64-bit if they want

Patch-set: 2