Update patch set 1
Patch Set 1:

i think if you want to make that distinction you need to be more explicit. i definitely hadn't got that, and i was _trying_ to work out if there was a reason! i think we'd be much better off with "prevented read past end of buffer" and "prevented write past end of buffer"!

Patch-set: 1