Update patch set 1
Patch Set 1:

what i mean is, if we move __thread_entry into the same .a, the linker will know that it can use a 32-bit relocation because it knows that __thread_entry can't be too far away at runtime, so the warning will go away and we won't have had to indirect. no?

Patch-set: 1