Correct the return value of setvbuf
When setting fully buffered or line buffered, if buf == NULL setvbuf
tries to allocate a buffer itself twice, fistly with the buffer size = size
(the fourth parameter) and then with the buffer size = iosize. if it succeeds
the second time, the request is be honored, so the value returned should be 0
but not EOF in this situation.

Change-Id: I981831973817cee0779b62304acdb110d11de80c