Fix up failing glibc tests.
For some reason running clone in the same process causes pthread
calls later in other tests to fail. Move that test to fork itself instead.

Change the expected output of the stdio test for a glibc difference.

Remove the pause test from unistd. A better version of the test is in
signal_test.cpp. The way the test works currently causes a later sigaction
of SIGALRM to fail.

After this, running bionic-unit-tests-glibc passes for all tests.

Change-Id: Ib304eae4164115835a54991dfdca5821ecc3db5e