Update patch set 3
Patch Set 3: Code-Review+2

(1 comment)

i still like the idea of linking to the bug so when we're looking at this in two years we can either say "oh, yeah, that got fixed and we can remove this hack" or "it wasn't that, maybe we need to investigate more".

oh, hang on... fixed in glibc 2.14? my z620's on 2.15 already.

Patch-set: 3
Label: Code-Review=+2