Update patch set 5
Patch Set 5:

The issue is that compiler-rt includes references to __aeabi_mem* functions, but anything linking against libc uses the libc implemented __aeabi_mem functions. This breaks runtime for these modules. Before this change, I was only building with clang except for libc.

There are two decent solutions, and one not so decent solution as far as I see it.

1) This patch.
2) Change link order in LLVM, remove these __aeabi_mem* functions completely from libc, fully link to libgcc, and suffer greatly from linked size increases.

3) Implement support in compiler-rt and clang (possibly only clang) for a flag such as -fno-builtin-aeabi-mem (this one should be self explanatory) that could be used when building these modules.

(3) may actually be the best solution here because compiler-rt's behavior is this way due to libgcc's behavior. However, we're not even linking libgcc that way due to the size increase (why code like this and all of libgcc_compat is even needed). I'll take a look at this change since it should be quick.

Thank you.

Patch-set: 5
Reviewer: Gerrit User 1022430 <1022430@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0