Add backtraces debug support to bionic mmap.
This change introduces the ability of recording backtraces for
anonymous mmap, improving the framework support to memory debugging.
From a conceptual viewpoint, this is complementing the
libc.debug.malloc support. The current malloc_leak implementation,
in fact, is not able to capture stack traces for: (a) large mallocs
which exceeded the MMAP_THRESHOLD; (b) direct calls to mmap (e.g.,
as the V8 engine is used to do).
From a technical viewpoint, this change introduces a new property,
named libc.debug.mmap. When this is set to 1, the stack trace for
each mmap call is collected and set as name of the mapped region.
This is possible thanks to the BIONIC_PR_SET_VMA_ANON_NAME prctl
which is available on Android kernels.
The backtracke string is therefore available in the
/proc/pid/(s)maps interface and in showmap -a.
For instance, the output of /proc/xxx/maps changes from:
b6f44000-b6f45000 r--p 00000000 00:00 0 [anon:libc_malloc]
to:
b6f44000-b6f45000 r--p 00000000 00:00 0 [anon:b6f7b768 b6f979e0...]
when libc.debug.mmap = 1.
Furthermore, this change is introducing a a python script for
symbolizing the stacktraces.

Change-Id: Idd3591167be44884a9cf1f9f73df032fccddecc1