Linker: Make linker parse all files in LD_LIBRARY_PATH until find a proper one
In current linker implementation, when LD_LIBRARY_PATH is set, linker will try to load
the first library, which matches the library name. If this library doesn't fit, linker
will return NULL and igore the remaining PATH.

This behaviour could be wrong if we have both 32bit libraries and 64bit ones in the system.
This patch will makes linker be able to parse all the paths in LD_LIBRARY_PATH until it
finds a proper library.

Change-Id: Ib03660c555e61e34ce2d98c93d524b6cc304d1dd
Signed-off-by: Qiming Shi <qiming.shi@intel.com>
Author-tracking-issue: ABIT-28