Update patch set 1
Patch Set 1: Code-Review+1

Change LGTM, but I can't help but notice that res_searchN in getaddrinfo.c and res_nsearch in res_query.c (and similarly res_querydomainN and res_nquerydomain) are almost identical.

The only differences seem to be in how they set h_errno and the fact that the getaddrinfo copy stores the data in struct res_target instead of in function parameters because it needs to store more than one answer.

Is it worth trying to get rid of one of them?

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1