Update patch set 2
Patch Set 2:

> actually, there's something weird here. the declaration inside the
 > extern "C" {} in fcntl.h (which is included in open.cpp) should
 > mean you don't need to repeat yourself here. is this a clang bug?

I would have to look at the C++ standard to see, but I doubt that this is a bug (or there would be more serious issues). I think the problem is that you are defining this function in a C++ source file. Just because you declared a "C" version probably doesn't mean that this definition should be used to fill it.

Patch-set: 2
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0