Update patch set 1
Patch Set 1:

well, that's the sad thing about these fortify functions and why we need this patch at all --- they are ABI, and we're stuck supporting them forever.

we could move them inside the fortify ifdefs (but not the clang ones). that would be exactly like glibc. this stdio.h change is already like that, and so were all the files i didn't touch, so i'll change fcntl.h to match.

Patch-set: 1
Label: Verified=0