Update patch set 1
Patch Set 1:

where did you see this? looking at the git history, it does look like this is what the old struct used to be, so i'm inclined to believe we should take this patch, but this is the first we've heard (afaik) of any trouble, which is surprising if it's wrong.

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0