Update patch set 1
Patch Set 1:

It looks like this would cause __system_property_find_compat to read 4 garbage entries that should be reserved before getting to the real toc, which means count will be off by 4 and it will never find the last 4 properties.  We could easily have failed to catch that in our testing.

Patch-set: 1