Update patch set 1
Patch Set 1:

i don't _think_ any of them are Android-specific hacks, though, are they? the ones i see look like projects that just choose to use their own netinet/tcp.h. (note that the linux kernel structures we expose are very different from the regular userspace ones. i wonder if any calling code is actually happy with what we offer. would we be better off just replacing this whole file with the BSD one? [i'm assuming that none of the BSD constants are different; we're safe when a header file only contains on-the-wire stuff, but not safe when there's a kernel interface involved. i _think_ netinet/tcp.h is only on-the-wire stuff defined in RFCs, but i'm not 100% certain of that.])

Patch-set: 1