Update patch set 1
Patch Set 1: Code-Review-2

I think we need to check this in res_send.c  res_nsend so we treat an error from a dns server as a signal to try another rather than a successful negative-cache situation.

I'm not too concerned about it though - these errors signify a meltdown on the server or the use of a questionable server and shouldn't be seen in practice.  I'll create a bug to look at this in the future in case the sony guys are not working on this anymore.

Patch-set: 1
Reviewer: Gerrit User 1002609 <1002609@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2