Update patch set 3
Patch Set 3:

(3 comments)

Sorry should have marked this WIP since there are a couple of other changes I need to make.

I couldn't use the BEGIN idea because I couldn't figure out a way to actually make it work. If we had something like unifdef I could do something clever, but a macro can't include #if or #else. Adding those macros only seemed to make things a bit more confusing since the end macro was always nothing.

I thought of them becoming something like if (false) { but I don't think this guarantees the code in the function body won't compile and a lot of the code simply won't compile on the host.

I also thought about creating a file that contained cruft to allow this stuff to compile, but that seemed like a bit of a rat hole. So I left it like this.

Patch-set: 3
Label: Verified=0